Clean up filtering conditions in Penthesilea.
The filtering conditions in the SiPM has been grouped as a filtering
function and moved to s12_filter.py in filters module.
